mruby-compiler: compare integers of the same signedness
The `cxx_abi` build compiles the compiler sources as C++, where gcc's
`-Wall` implies `-Wsign-compare`; the C builds would need `-Wextra` for
it. Four comparisons therefore warn in that build alone, and every gcc
job in CI prints each of them twice, once for the `mrbc` sub-build and
once for the library.

`ccontext.c` walked a `size_t` count with an `int` index, while the
loop over the same `locals_count` further down the file already uses
`size_t`. `compile.c` kept `capacity` as `int` next to a `size_t`
`length` and a `mrc_realloc` size argument. `cdump.c` held
`MRC_STRING_LEN`, a `size_t`, in an `mrc_int`. `codegen.c` compared
`data.a`, a `uint32_t`, against `dst+1`, where the `uint16_t` `dst` is
promoted to `int`.

+++ mrbgems/mruby-compiler/src/codegen.c
@@ -1438,7 +1438,7 @@ gen_binop(mrc_codegen_scope *s, mrc_sym op, uint16_t dst)
   else if (op == MRC_OPSYM_2(aref)) {
     /* GETIDX0 fusion: MOVE dst arr; LOADI_0 dst+1 -> GETIDX0 dst arr */
     struct mrc_insn_data data = mrc_last_insn(s);
-    if (data.insn == OP_LOADI_0 && data.a == dst+1 && addr_pc(s, data.addr) != s->lastlabel) {
+    if (data.insn == OP_LOADI_0 && data.a == (uint32_t)(dst+1) && addr_pc(s, data.addr) != s->lastlabel) {
       struct mrc_insn_data data0 = mrc_decode_insn(mrc_prev_pc(s, data.addr));
       if (data0.insn == OP_MOVE && data0.a == dst && data0.b != dst) {
         s->pc = addr_pc(s, data0.addr);
